Add pins showing in SummaryActivity
